+++ CHANGELOG.md
@@ -3,10 +3,15 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-* New CMS functionality
+## 1.3.0 / 2017-12-04
+
+* New CMS functionality (#22)
 * Add Portuguese as an admin language
 * Add preset color palettes
 * Require Rails 5.1.0 or above
+* Fix issue with seeding content (#38)
+* Add Italian translations
+* Use property-web-scraper to import from web pages
 
 ## 1.2.0 / 2017-08-23